+++ src/commonMain/kotlin/kotools/types/experimental/NonZeroInt.kt
@@ -9,6 +9,7 @@ import kotools.types.internal.ExperimentalSince
 import kotools.types.internal.KotoolsTypesVersion
 import kotools.types.internal.unexpectedCreationError
 import kotools.types.number.NonZeroInt
+import kotools.types.number.StrictlyNegativeInt
 import kotools.types.number.StrictlyPositiveInt
 import kotools.types.number.toNonZeroInt
 import kotlin.jvm.JvmSynthetic
@@ -37,6 +38,24 @@ public operator fun NonZeroInt.unaryMinus(): NonZeroInt {
         .getOrNull()
         ?: unexpectedCreationError<NonZeroInt>(value)
 }
+
+/**
+ * The negative range of values a [NonZeroInt] can have.
+ *
+ * Here's an example of calling this property from Kotlin code:
+ *
+ * ```kotlin
+ * println(NonZeroInt.negativeRange) // [-2147483648;-1]
+ * ```
+ *
+ * Please note that this property is not available yet for Java users.
+ */
+@ExperimentalKotoolsTypesApi
+@ExperimentalSince(KotoolsTypesVersion.Unreleased)
+@get:JvmSynthetic
+public val NonZeroInt.Companion.negativeRange:
+        NotEmptyRange<StrictlyNegativeInt>
+    get() = StrictlyNegativeInt.range
 
 /**
  * The range of positive values a [NonZeroInt] can have.