+++ src/commonTest/kotlin/kotools/types/experimental/NonZeroIntTest.kt
@@ -3,18 +3,20 @@
  * Use of this source code is governed by the MIT license.
  */
 
+
 package kotools.types.experimental
 
 import kotools.types.internal.unexpectedCreationFailure
 import kotools.types.number.NonZeroInt
+import kotools.types.number.StrictlyNegativeInt
 import kotools.types.number.StrictlyPositiveInt
 import kotools.types.number.toNonZeroInt
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertSame
 
-@ExperimentalKotoolsTypesApi
 class NonZeroIntTest {
+    @OptIn(ExperimentalKotoolsTypesApi::class)
     @Test
     fun unaryMinus_should_pass() {
         val number: NonZeroInt = NonZeroInt.random()
@@ -26,8 +28,15 @@ class NonZeroIntTest {
     }
 }
 
-@ExperimentalKotoolsTypesApi
+@OptIn(ExperimentalKotoolsTypesApi::class)
 class NonZeroIntCompanionTest {
+    @Test
+    fun negativeRange_should_be_StrictlyNegativeInt_range() {
+        val actual: NotEmptyRange<NonZeroInt> = NonZeroInt.negativeRange
+        val expected: NotEmptyRange<NonZeroInt> = StrictlyNegativeInt.range
+        assertSame(expected, actual)
+    }
+
     @Test
     fun positiveRange_should_be_StrictlyPositiveInt_range() {
         val actual: NotEmptyRange<NonZeroInt> = NonZeroInt.positiveRange